Basic movement, anim setup, and door setup
The player has no acceleration or sprint or anything its just basic rigidbody velocity. Since we're on the newest version of unity we are using the "New" input system, its going to be a lot nicer in the long run but it might take some getting used to.
The anim stuff is obviously just temporary, we will change out the sprites in the anim controller. Wasn't sure if idle should be directional or not so just made it default to look up.
The doors work with an ID system, each pair of doors that will link should just use the same ID and everything should work.

IMPORTANT: --- The Scene "AlwaysActive" Must Always Be Active, or else the doors will not work (or most other things probably). If its not in your hierarchy just drag in there while another scene is loaded, both can (and should) be active at once. ---