started work on NoteBook
also fixed door not setting the right scene to be active/where the player technically "is" once loaded. Not sure whether or not it really maters what scene is the currently loaded one but I think it should work better this way.
started work on disabling inputs when in the notebook but not done with it yet