added multiple pages for notes
can be cycled through with either left/right arrows, or 1 / 2 number keys. (will change the keybinds to be more sensible later). currently there is only 3 pages but it is possible to add / remove an arbitrary amount and it will still work (we will likely have 16, one for each character + a free page)