camera follow player with bounds
I decided to just use cinemachine for this as its a lot easier than trying to make something manually. Also change how player "spawning" works, now the player is just always active on scene load and is simply moved to where they need to be. This will make it a lot easier to do stuff to player scripts so I think its probably worth doing it this way instead of spawning from the prefab